+++ models/device_twin.go
@@ -1,6 +1,7 @@
 package models
 
 import (
+	"context"
 	"github.com/thingio/edge-device-std/logger"
 )
 
@@ -11,27 +12,21 @@ type DeviceTwin interface {
 	// It must always return nil if the device needn't be initialized.
 	Initialize(lg *logger.Logger) error
 
-	// Start will to try to create connection with the real device.
+	// Start will to try to open a connection with the real device.
 	// It must always return nil if the device needn't be initialized.
-	Start() error
-	// Stop will to try to destroy connection with the real device.
+	Start(ctx context.Context) error
+	// Stop will to try to close a connection with the real device.
 	// It must always return nil if the device needn't be initialized.
 	Stop(force bool) error
-	// HealthCheck is used to bin the connectivity with the real device.
+	// HealthCheck is used to check the connectivity with the real device.
 	HealthCheck() (*DeviceStatus, error)
 
-	// Watch will read device's properties periodically with the specified policy.
-	Watch(bus chan<- *DeviceDataWrapper) error
-	// Read indicates soft read, it will read the specified property from the cache with TTL.
-	// Specially, when propertyID is "*", it indicates read all properties.
+	// Read indicates hard read, it will read the specified property from the real device.
 	Read(propertyID ProductPropertyID) (map[ProductPropertyID]*DeviceData, error)
-	// HardRead indicates head read, it will read the specified property from the real device.
-	// Specially, when propertyID is "*", it indicates read all properties.
-	HardRead(propertyID ProductPropertyID) (map[ProductPropertyID]*DeviceData, error)
 	// Write will write the specified property to the real device.
 	Write(propertyID ProductPropertyID, values map[ProductPropertyID]*DeviceData) error
 	// Subscribe will subscribe the specified event,
-	// and put DataOperation including properties specified by the event into the bus.
+	// and you should put the event into the bus.
 	Subscribe(eventID ProductEventID, bus chan<- *DeviceDataWrapper) error
 	// Call is used to call the specified method defined in product,
 	// then waiting for a while to receive its response.